create ManifestValidator for reuse
